Improve SDS download code
Just after an ACK or NAK, a WAIT packet for the next message is sent and queued in the device.
This ensures that the device will not send more packets until it receives an ACK or a NAK and
fixes some corrupted messages caused when trying to recover from errors (the corruption
apparently happened when the device was receiving data while sending the data packets).
Besides, the code has been reworked to be more reliable against incomplete packets (timeout)
by retrying the download.