Improve SDS handshake
As the handshake implementation was only tested on an E-Mu ESI-2000, it
presented some issues when run over other devices. In particular, the
only way the E-Mu would reply safely to a message was to send an illegal
sample dump message but this is not working with an Elektron Machinedrum.

Other handshake methods have been implemented by using SDS messages and
the original implementation is only used as a fallback now.