Use non blocking inotifier
This approach allows to process batch of events, which reduces the
scheduled calls to browser_load_dir. Notice that the browser
implementation skips calls to browser_load_dir if one scheduled call
was already running and if two events were too close in time only the
first call was actually refreshing the view.