Remove slot_used member in item struct
It was used to indicate if a slot was used but this feature was only
used in the data filesystems in the elektron connector and only as a
fallback in a couple of scenarios.
The potential issue was that slots would not be shown at all in the
GUI leading to gaps in the slot sequence and ultimately not allowing
to drop elements on them.