Fix occasional failure when loading samples that need resampling
Regression added in 5a5d6019c555f575a492dc3036fccfcbecd7fecb for samples
that need resampling when the estimated amount of frames is not the actual
amount returned by the resampling library at the end of the process.

If a rounding fix is needed, the callback needs to be called to properly
notifiy the end of the loading process.