Speed up Elektron package downloading
After inroducing .metadata files reading when iterating over DATA filesystems
in 9d98084158436abda956297583bd07ec0ac1d777, the whole reading could take too
much. In particular, reading a directory is the only way of getting the name
of a DATA item, which is used several times when downloading a package.
With this little hack we can control the .metadata load.