Speed up handshake
Now, connectors declare if their handshake requires the standard MIDI handshake.
The idea is to avoid duplicate calls to the standard MIDI handshake as the
timeout can cause a noticiable delay.

If a connector name is used, it is guaranteed that it will be called if needed.
If not, it is only called once for all the connectors.