Use item id to avoid repeated calls when selecting items.
From now on, the member id in the item struct must always be filled up
with a number that identifies the item. For standard (non slot) filesystems,
it is enough to be a sequential number for the current directory

This is a rework of 4fbe512f5b5389701536a59d2389412f6adf0471.