Change get_exts function signature in fs_operations struct
This is a preparation for reading the Elektron filesystem extensions from the JSON
file. As this is a quite complicated refactor, changes in the application and all the
connectors except the Elektron one, are done in this commit.
All Elektron connectors have been disabled in this commit and will be enabled again
in the next commit.