Allow floating point as the internal representation for samples
Elektroid tries its best to not get in the middle when editing samples by saving
the samples in the same format they were originally loaded from.

The problem is that when using the editor, since Elektroid uses a 16 bit signed
integer for the internal audio representation, when saving back to formats that
might have greater resolution, this extra resolution is lost. For instance, a 24
bit signed integer will loose the additional 8 bits when saving as they will appear
as zeroes.

On top of this, Elektroid purpose is to be fast and require as lower resources as
possible. Using floating point audio data might increase some overhead when playing.

Due to all this, Elektroid allows the user to choose between using floating point
or 16 bit signed integer.

Note that conversion issues only appear when using the editor. Sample conversions when
uploading or downloading do not have this issue.