Comment another failing test
A test fails on Digitakt I with versions newer than 1.40B because of a firmware
bug. The code has been kept for hitorical reasons but it makes no sense to keep
testing this case as this is unlikely to be fixed.

bc5368bb30e2df7cdae14589c34e903a349c28f2 fixed the same error on another test.