Fix loop selection on some samples
In some cases, libsamplerate generates an additional frame in stereo while it
does not do it in mono. The code was not able to handle this additional frame,
causing a mismatch between the frames and the size of the array, which is used
to detect the end of the loading process, and ultimately causing the clicks on
the editor to be ignored.
To fix this, only when the actual amount of frames is less than the previously
calculated upper bound for the amount of frames, the frame amount is updated;
if not, the additional generated frames can be ignored safely.

This was kind of introduced in 0282cd2b920fc2b183d2d6e036f0c9ce53b739e6, as no
additionals checks were performed.