GUI: Skip showing progress window when unneeded
The solution was already in use when deleting items. Now, this solution
has been brought to uploads and downloads.

The idea is to skip this if the source browser is of type SYSTEM, which
makes the action instantaneous, but there is a limit of items too to
prevent long waits without an indication.