GUI: Improve notifier
Under some circumstances, there were too many events observed and the
GUI was lagging and blocking the browser. Only delete and create events
are observed now. Still, some events might be lost.

Parent dir notifier has been removed as it is enough with the current
dir notifier.

When a dir is deleted, it is not easy to determine which should be the
next dir to be in. The parent dir might have been deleted too but the
notification might have been skipped so it is safer to move to the user
home.