GUI: Fix issue with label in progress window
As the label text was stored and set when refreshing, the text needed
to be statically allocated. This was an issue when showing the window
after a dinamically allocated text was set.
This commit sets the text at the moment, does not update it when
refreshing and does not store it whatsoever.