Add support for sample tags
Sample tags are stored in the IKEY chunk of the LIST INFO chunk.
Only filesystems using the FS_OPTION_SHOW_SAMPLE_COLUMNS option require
to fill the sample_info struct in the item struct.

When loading a sample_info struct, the tags member of the sample_info
struct might be set to NULL. This is more efficient when iterating in
the browser.
When loading a sample, the tags member is required to be initialized
just like any other member of the sample_info struct.

Keep in mind that files different that WAV will require exporting as saving
can not guarantee that the tags will be saved along with the data.

Also, saving the tags is transparent. All the tags read are kept when
writing and LIST INFO will not be written if there are no read tags.