Add warning when using RtMidi on Linux
Some devices require long SysEx messages and this can not be changed.

In RtMidi, due to its architecture, buffers need to be as long as these
MIDI messages and this requires a long ALSA MIDI sequencer ringbuffer
to prevent buffer overruns.

This is not needed when using the ALSA rawmidi API.