GUI: Change how editor saves edited samples
There was an issues when saving samples in other formats than WAV when
honouring the sample_info_src member in the audio struct as not all the
formats allow loop points of tags, which is an upcoming feature.

The new approach is saving samples using the value in sample_info_src
but forcing the WAV format so that sample rate and bitdepth are kept.

In order to let the user know about this change while keeping things as
transparent as possible, besides the save menu button there is a new
export menu button but both are mutually exclusive from the user POV.
Both run the same code though.